Load assets in respective states.
It doesn't make sense to load all assets in a single state (Splash state in
our case). This is because each state will have it's own set of assets and
we don't want them to clash. A cleaner approach would be to simply load the
assets in their respective states using the preload function.