Extract game dimension variables into config.
The two games we have are of different sizes, one runs on portrait whilst
the other on landscape. We need to dynamically change our game canvas size
when we transition from one game to another. This commit introduces two new
objects in config which keeps track of the game sizes.