Load the collectible coins image.
Phaser does annimations using a keyframe-based approach which means every x
seconds, phaser changes the asset such that it looks like the sprite is
animated. We can specify all the various forms of a sprite in a single png
and then load this png using the `load.spritesheet` method by specifying
the dimensions of each form.