Allow hero to collect coins.
This is another collision detection between our hero and coins. One
difference here (vs. platforms) is that we use the `overlap` method instead
of the `collide` since we simply want to know if the hero touched a coin
but still allow hero to move through the coin (collide prevents sprites to
pass through each other).