Spawn invisible walls.
We want the spiders to crawl only on the platforms they are spawned upon.
We can do this by rendering invisible walls into our game and then
detecting collisions with spiders. Our hero will be unaffected by these
walls and will be able to pass through them.

Since we want an invisible wall on either side of a platform, we can
call a method to create them from our `_spawnPlatforms` method.