Make the spiders turn.
Upon collision with invisible walls, we want the spiders to turn. We use
the `touching` along with the `blocked` methods to do this. Note that we
require to use the `blocked` method in addition to `touching` because we
want the spiders to turn if they collide with the world bounds as well.